Add alien and butterfly drawables, drawing actions and viewmodel
This commit introduces two new drawable resources, `alien.xml` and `butterfly.xml`, which can be used to display vector graphics in the app.

It also adds `DrawingAction.kt` which defines the possible actions a user can take when drawing, such as starting a new path, drawing, ending a path, undoing/redoing, clearing the canvas, and selecting a color.

The `DrawingViewModel.kt` manages the state of the drawing canvas and handles the different drawing actions.