Adjust StatisticsScreen and add Outfit font
Updates the StatisticsScreen to add horizontal padding to the column containing the statistics items for better layout.

Adds the Outfit font family with Medium, Regular, and SemiBold variants. Updates the typography styles (`bodyLarge`, `bodyMedium`, `bodySmall`, `labelLarge`, `labelMedium`, `labelSmall`, `headlineLarge`) to use the Outfit font family and adjusts font weights.

Modifies the StatisticsItem composable to give the title Text a weight of 1f to occupy available space and sets a fixed width and end alignment for the percentage Text for consistent layout.