Refactor: Enhance feedback screens and display components
Modifies the `FinalFeedbackScreen` to accept a `gameType` parameter to display different messages for Speed Draw and Endless Mode.
Updates `DrawingGraph.kt` to pass the `gameType` to `FinalFeedbackScreen`.

The `SpeedDrawFeedbackCard` and `DisplayCounter` composables now accept a `shouldShowHighScore` boolean.
`SpeedDrawFeedbackCard` will conditionally display the `HighScoreBadge` based on this new parameter.
`DisplayCounter` will show "New High" text when `shouldShowHighScore` is true.

Updates usage of `DisplayCounter` in `SpeedDrawScreen` and `EndlessModeScreen` to pass `shouldShowHighScore = false`.
Updates usage of `SpeedDrawFeedbackCard` in `FinalFeedbackScreen` to pass `shouldShowHighScore = false`.